requirements.txt
# Require Python 3.11 or 3.12
dora-rs
#torch
#torchvision
#torchaudio
opencv-python
# YOLOv5 requirements
# Usage: pip install -r requirements.txt

# Base ----------------------------------------

javascript>=1!1.1.1
selenium>=4.15.2
matplotlib>=3.2.2
numpy>=1.18.5
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
--extra-index-url https://download.pytorch.org/whl/cpu
tqdm>=4.64.0
ultralytics>=8.0.238
protobuf<=3.20.1  # https://github.com/ultralytics/yolov5/issues/8012
openai-whisper
sounddevice>=0.4.1
OpenAi
# Logging -------------------------------------
tensorboard>=2.4.1
# wandb
# clearml

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Export --------------------------------------
# coremltools>=5.2  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.4.1  # ONNX simplifier
# nvidia-pyindex  # TensorRT export
# nvidia-tensorrt  # TensorRT export 
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export (or tensorflow-cpu, tensorflow-aarch64)
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop>=0.1.1  # FLOPs computation
# albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
# roboflow

opencv-python>=4.1.1
pyarrow
maturin

#npm i minecrafthawkeye